Create run directory module
